🚚 rename and move modules
