+++ src/laser_measles/__init__.py
@@ -1,9 +1,25 @@
 __version__ = "0.0.1"
 
+from .antibodies import MaternalAntibodies
+from .births import Births
 from .core import compute
+from .incubation import Incubation
+from .infection import Infection
 from .model import Model
+from .mortality import NonDiseaseDeaths
+from .routine import RoutineImmunization
+from .susceptibility import Susceptibility
+from .transmission import Transmission
 
 __all__ = [
     "compute",
+    "Births",
+    "Incubation",
+    "Infection",
+    "MaternalAntibodies",
     "Model",
+    "NonDiseaseDeaths",
+    "RoutineImmunization",
+    "Susceptibility",
+    "Transmission",
 ]
