+++ src/laser_measles/generic/__init__.py
@@ -1,0 +1,4 @@
+from .metapop import get_scenario
+from .params import get_parameters
+
+__all__ = ["get_scenario", "get_parameters"]
