+++ src/laser_measles/incubation.py
