+++ src/laser_measles/infection.py
