+++ src/laser_measles/nigeria/nga_adm_osgof_20190417/nga_admbnda_adm2_osgof_20190417.shp.xml
